Refactoring based on example of gitlabhq project
